refactor(buddy): move pure geometry/dock to src/shared for renderer reuse — old paths re-export
