+++ desktop/src/main/buddy-window-manager.ts
@@ -9,6 +9,12 @@ import {
   dockReducer, detectSnapEdge, dockPosition, FREE_DOCK, SNAP_THRESHOLD_PX,
   type DockState, type DockEvent, type DockEdge,
 } from './buddy-dock';
+// Rect/Point/Size/clampToWorkArea used to be DEFINED in this file; they now
+// live in src/shared/buddy-geometry.ts so the renderer overlay window can use
+// them too (renderer code can never import from src/main/, which pulls in
+// electron). Imported here for this file's own use, and re-exported below so
+// every existing `from './buddy-window-manager'` import keeps working.
+import { clampToWorkArea, type Rect, type Point, type Size } from '../shared/buddy-geometry';
 
 // Push-channel names (kept as local consts — this module deliberately doesn't
 // import shared/types; values must match IPC.* in src/shared/types.ts).
@@ -22,22 +28,8 @@ const IPC_MASCOT_STATE = 'buddy:mascot-state';
 // ~17px off the left edge, so a 24px test would fire on ordinary positioning.
 const PEEK_PAST_EDGE_PX = 6;
 
-export interface Rect { x: number; y: number; width: number; height: number; }
-export interface Point { x: number; y: number; }
-export interface Size { width: number; height: number; }
-
-/**
- * Clamp a position so the window stays fully inside the workArea.
- * Pure function — no electron deps — so it's unit-testable.
- */
-export function clampToWorkArea(pos: Point, size: Size, workArea: Rect): Point {
-  const maxX = workArea.x + workArea.width - size.width;
-  const maxY = workArea.y + workArea.height - size.height;
-  return {
-    x: Math.max(workArea.x, Math.min(pos.x, maxX)),
-    y: Math.max(workArea.y, Math.min(pos.y, maxY)),
-  };
-}
+export { clampToWorkArea } from '../shared/buddy-geometry';
+export type { Rect, Point, Size } from '../shared/buddy-geometry';
 
 // Window sizes + all position math live in buddy-bar-geometry.ts (pure,
 // unit-tested); main.ts imports the same constants so the BrowserWindow
